breaking: rename current steps module into a legacy module
