+++ mxops/execution/scene.py
@@ -13,7 +13,7 @@ import yaml
 
 from mxops.config.config import Config
 from mxops.data.execution_data import _ScenarioData, ExternalContractData, ScenarioData
-from mxops.execution.steps import LoopStep, SceneStep, Step, instanciate_steps
+from mxops.execution.legacy_steps import LoopStep, SceneStep, Step, instanciate_steps
 from mxops.execution.account import AccountsManager
 from mxops import errors
 from mxops.utils.logger import get_logger
